feat: Apollo gateway
* Setup an Apollo gateway to route to Hasura at /api/v1/graphql
* Verified that the Apollo sandbox works with this
* Setup a local subgraph at /api/v1/graphqlSubgraph, for future use
* Update Hasura3 config to support Apollo federation
