+++ .github/workflows/ci-default.yml
@@ -11,6 +11,7 @@ env:
   PLASMIC_PROJECT_API_TOKEN: ${{ vars.PLASMIC_PROJECT_API_TOKEN }}
   NEXT_PUBLIC_DOMAIN: "www.opensource.observer"
   NEXT_PUBLIC_DB_GRAPHQL_URL: ${{ vars.NEXT_PUBLIC_DB_GRAPHQL_URL }}
+  HASURA_URL: ${{ vars.HASURA_URL }}
   OSO_API_KEY: ${{ secrets.OSO_API_KEY }}
   NEXT_PUBLIC_ALGOLIA_APPLICATION_ID: "test"
   NEXT_PUBLIC_ALGOLIA_API_KEY: "test"
@@ -80,6 +81,7 @@ jobs:
             PLASMIC_PROJECT_API_TOKEN \
             NEXT_PUBLIC_DOMAIN \
             NEXT_PUBLIC_DB_GRAPHQL_URL \
+            HASURA_URL \
             OSO_API_KEY \
             GOOGLE_PROJECT_ID \
             GOOGLE_TEST_DUMMY_CREDENTIALS_JSON \
