+++ .github/workflows/refresh-test-credentials.yml
@@ -83,6 +83,7 @@ jobs:
       PLASMIC_PROJECT_API_TOKEN: ${{ vars.PLASMIC_PROJECT_API_TOKEN }}
       NEXT_PUBLIC_DOMAIN: ${{ vars.NEXT_PUBLIC_DOMAIN }}
       NEXT_PUBLIC_DB_GRAPHQL_URL: ${{ vars.NEXT_PUBLIC_DB_GRAPHQL_URL }}
+      HASURA_URL: ${{ vars.HASURA_URL }}
       OSO_API_KEY: ${{ secrets.OSO_API_KEY }}
       GOOGLE_PROJECT_ID: "opensource-observer"
       GOOGLE_TEST_DUMMY_CREDENTIALS_JSON: ${{ vars.GOOGLE_TEST_DUMMY_CREDENTIALS_JSON }}
@@ -108,6 +109,7 @@ jobs:
             PLASMIC_PROJECT_API_TOKEN \
             NEXT_PUBLIC_DOMAIN \
             NEXT_PUBLIC_DB_GRAPHQL_URL \
+            HASURA_URL \
             OSO_API_KEY \
             GOOGLE_PROJECT_ID \
             GOOGLE_TEST_DUMMY_CREDENTIALS_JSON \
