+++ apps/frontend/.env.local.example
@@ -15,6 +15,7 @@ PLASMIC_PROJECT_API_TOKEN=
 NEXT_PUBLIC_DOMAIN=www.opensource.observer
 ### GraphQL API endpoint for database (required)
 NEXT_PUBLIC_DB_GRAPHQL_URL=
+HASURA_URL=
 ### OSO API key
 OSO_API_KEY=
 ### Supabase for database (required)
