+++ apps/frontend/app/api/v1/graphqlSubgraph/route.ts
@@ -1,0 +1,36 @@
+import { ApolloServer } from "@apollo/server";
+import { buildSubgraphSchema } from "@apollo/subgraph";
+import { startServerAndCreateNextHandler } from "@as-integrations/next";
+import { NextRequest } from "next/server";
+import { gql } from "graphql-tag";
+
+const resolvers = {
+  Query: {
+    hello: () => "world",
+  },
+};
+
+const typeDefs = gql`
+  # highlight-start
+  extend schema
+    @link(
+      url: "https://specs.apollo.dev/federation/v2.0"
+      import: ["@key", "@shareable"]
+    )
+  # highlight-end
+
+  type Query {
+    hello: String
+  }
+`;
+
+const server = new ApolloServer({
+  schema: buildSubgraphSchema({ typeDefs, resolvers }),
+  introspection: true,
+});
+
+const handler = startServerAndCreateNextHandler<NextRequest>(server, {
+  context: async (req) => ({ req }),
+});
+
+export { handler as GET, handler as POST };
