+++ apps/frontend/lib/config.ts
@@ -25,6 +25,7 @@ export const DB_GRAPHQL_URL = requireEnv(
   process.env.NEXT_PUBLIC_DB_GRAPHQL_URL,
   "NEXT_PUBLIC_DB_GRAPHQL_URL",
 );
+export const HASURA_URL = process.env.HASURA_URL;
 export const OSO_API_KEY = process.env.OSO_API_KEY;
 export const SUPABASE_URL = requireEnv(
   process.env.NEXT_PUBLIC_SUPABASE_URL,
